Add an optional configuration( 'deploy.public_dir' ) for deployer
Instead of using the value of "public_dir" in the main configuration file, we can specify the folder for deployment (eg: public / blog /) via the value of "deploy.public_dir".
If we do not specify the value of "deploy.public_dir", the value of "public_dir" in the main configuration file will be used by default.